Fix Inventory Unit Mapping
The orphan-removal flag is on the wrong side of the association and is discarded by doctrine.
This leads to Order::removeInventoryUnit() failing to cascade properly, and makes the cart update fail when reducing the quantity of any item in the cart. Marketing would love to keep this bug, I'm sure ;)